+++ cmd/terramate/e2etests/cloud/run_script_cloud_sync_preview_test.go
@@ -58,7 +58,7 @@ func TestScriptRunWithCloudSyncPreview(t *testing.T) {
 
 	for _, tc := range []testcase{
 		{
-			name: "basic success sync",
+			name: "warning when not running with GITHUB_ACTIONS",
 			layout: []string{
 				"s:stack:id=someid",
 				`f:terramate.tm:
@@ -92,6 +92,63 @@ func TestScriptRunWithCloudSyncPreview(t *testing.T) {
 				"run:stack:terraform init",
 			},
 			cmd:             []string{"script", "run", "-X", "preview"},
+			githubEventPath: datapath(t, "interop/testdata/event_pull_request.json"),
+			want: want{
+				run: RunExpected{
+					Status: 0,
+					StdoutRegexes: []string{
+						"Plan: 1 to add, 0 to change, 0 to destroy.",
+					},
+					StderrRegexes: []string{
+						"--cloud-sync-preview is only supported in GitHub Actions workflows",
+					},
+				},
+				ignoreTypes: cmpopts.IgnoreTypes(
+					cloud.CommandLogs{},
+					&cloud.ChangesetDetails{},
+					cloudstore.Stack{},
+					&cloud.DeploymentMetadata{},
+				),
+			},
+		},
+		{
+			name: "basic success sync",
+			layout: []string{
+				"s:stack:id=someid",
+				`f:terramate.tm:
+				  terramate {
+					config {
+						experiments = ["scripts"]
+					}
+				  }
+				`,
+				fmt.Sprintf(`f:stack/preview.tm:
+
+				  script "preview" {
+					description = "sync a preview"
+					job {
+					  commands = [
+						["%s", "plan", "-out=out.tfplan", "-no-color", "-detailed-exitcode", {
+						  cloud_sync_preview             = true,
+						  cloud_sync_terraform_plan_file = "out.tfplan",
+						}],
+					  ]
+					}
+				  }
+				  `,
+					TerraformTestPath,
+				),
+				`f:stack/main.tf:
+				  resource "local_file" "foo" {
+					content  = "test content"
+					filename = "${path.module}/foo.bar"
+				  }`,
+				"run:stack:terraform init",
+			},
+			cmd: []string{"script", "run", "-X", "preview"},
+			env: []string{
+				"GITHUB_ACTIONS=1",
+			},
 			githubEventPath: datapath(t, "interop/testdata/event_pull_request.json"),
 			want: want{
 				run: RunExpected{
@@ -179,6 +236,7 @@ func TestScriptRunWithCloudSyncPreview(t *testing.T) {
 			s.Git().CommitAll("all stacks committed")
 
 			env := RemoveEnv(os.Environ(), "CI")
+			env = RemoveEnv(env, "GITHUB_ACTIONS")
 			env = append(env, "TMC_API_URL=http://"+addr)
 			env = append(env, "TM_GITHUB_API_URL=http://"+addr+"/")
 			env = append(env, "GITHUB_EVENT_PATH="+tc.githubEventPath)
@@ -191,44 +249,45 @@ func TestScriptRunWithCloudSyncPreview(t *testing.T) {
 			result := cli.Run(tc.cmd...)
 			AssertRunResult(t, result, tc.want.run)
 
-			orguuid := string(cloudData.MustOrgByName("terramate").UUID)
-			req, err := http.NewRequest("GET", "http://"+addr+"/v1/previews/"+orguuid+"/1", nil)
-			if err != nil {
-				t.Fatalf("failed to create request: %v", err)
-			}
-
-			req.Header.Set("User-Agent", "terramate/0.0.0-test")
-			httpClient := &http.Client{}
-			httpResp, err := httpClient.Do(req)
-			if err != nil {
-				t.Fatalf("failed to send request: %v", err)
-			}
-			defer func() { _ = httpResp.Body.Close() }()
-
-			assert.EqualInts(t, httpResp.StatusCode, http.StatusOK)
 			var previewResp cloudstore.Preview
-			if err := json.NewDecoder(httpResp.Body).Decode(&previewResp); err != nil {
-				t.Fatalf("failed to decode response: %v", err)
-			}
+			if tc.want.preview != nil {
+				orguuid := string(cloudData.MustOrgByName("terramate").UUID)
+				req, err := http.NewRequest("GET", "http://"+addr+"/v1/previews/"+orguuid+"/1", nil)
+				if err != nil {
+					t.Fatalf("failed to create request: %v", err)
+				}
 
-			for i, sp := range previewResp.StackPreviews {
-				assert.EqualStrings(t, tc.want.stackPreviewChangesets[i].Provisioner, sp.ChangesetDetails.Provisioner)
-				for _, asciiRegex := range tc.want.stackPreviewChangesets[i].ChangesetASCIIRegexes {
-					matched, err := regexp.MatchString(asciiRegex, sp.ChangesetDetails.ChangesetASCII)
-					assert.NoError(t, err, "failed to compile regex %q", asciiRegex)
+				req.Header.Set("User-Agent", "terramate/0.0.0-test")
+				httpClient := &http.Client{}
+				httpResp, err := httpClient.Do(req)
+				if err != nil {
+					t.Fatalf("failed to send request: %v", err)
+				}
+				defer func() { _ = httpResp.Body.Close() }()
 
-					if !matched {
-						t.Errorf("ChangesetASCII=\"%s\" does not match regex %q",
-							sp.ChangesetDetails.ChangesetASCII,
-							asciiRegex,
-						)
-					}
+				assert.EqualInts(t, http.StatusOK, httpResp.StatusCode)
+				if err := json.NewDecoder(httpResp.Body).Decode(&previewResp); err != nil {
+					t.Fatalf("failed to decode response: %v", err)
+				}
+				if diff := cmp.Diff(*(tc.want.preview), previewResp, tc.want.ignoreTypes); diff != "" {
+					t.Errorf("unexpected  preview: %s", diff)
 				}
 			}
 
-			if tc.want.preview != nil {
-				if diff := cmp.Diff(*(tc.want.preview), previewResp, tc.want.ignoreTypes); diff != "" {
-					t.Errorf("unexpected  preview: %s", diff)
+			if tc.want.stackPreviewChangesets != nil {
+				for i, sp := range previewResp.StackPreviews {
+					assert.EqualStrings(t, tc.want.stackPreviewChangesets[i].Provisioner, sp.ChangesetDetails.Provisioner)
+					for _, asciiRegex := range tc.want.stackPreviewChangesets[i].ChangesetASCIIRegexes {
+						matched, err := regexp.MatchString(asciiRegex, sp.ChangesetDetails.ChangesetASCII)
+						assert.NoError(t, err, "failed to compile regex %q", asciiRegex)
+
+						if !matched {
+							t.Errorf("ChangesetASCII=\"%s\" does not match regex %q",
+								sp.ChangesetDetails.ChangesetASCII,
+								asciiRegex,
+							)
+						}
+					}
 				}
 			}